refactor(flight): split leaf-prune + terminal-orbit/position into ParsekFlight.TerminalOrbit.cs
